// apps/web/src/api/generated/organizer-tournaments/upsertTournamentPaymentMethodDtoType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type UpsertTournamentPaymentMethodDtoType = typeof UpsertTournamentPaymentMethodDtoType[keyof typeof UpsertTournamentPaymentMethodDtoType];


export const UpsertTournamentPaymentMethodDtoType = {
  INSTAPAY: 'INSTAPAY',
  VODAFONE_CASH: 'VODAFONE_CASH',
  BANK_TRANSFER: 'BANK_TRANSFER',
  EXTERNAL_LINK: 'EXTERNAL_LINK',
  OTHER: 'OTHER',
} as const;