// apps/web/src/api/generated/tournamentPaymentMethodResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { TournamentPaymentMethodResponseDtoType } from './tournamentPaymentMethodResponseDtoType';

export interface TournamentPaymentMethodResponseDto {
  id: string;
  type: TournamentPaymentMethodResponseDtoType;
  displayName: string;
  enabled: boolean;
  /** @nullable */
  accountHolderName?: string | null;
  /** @nullable */
  accountIdentifier?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  instapayAddress?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankBranch?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  iban?: string | null;
  /** @nullable */
  swiftCode?: string | null;
  /** @nullable */
  externalUrl?: string | null;
  instructions: string;
  /** @nullable */
  notes?: string | null;
}